requirements.txt
# Add your requirements here like:
numpy>=1.3
biopython>=1.53
